# guangdong-energy-monitor/docs/API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: 广东省电厂监控平台 API 契约
  description: |
    华电集团广东电厂监控平台后端接口契约文档。
    所有接口统一通过 Gateway（8080）访问，需携带 JWT Token。
    所有响应统一使用 R<T> 包装结构。
    包含认证授权、用户管理、角色管理、组织架构、电厂管理、水文站管理、
    GIS 空间服务、报警管理、预警管理、规则管理、统计分析、预测服务、数据采集共13个控制器、54个REST端点。
  version: 1.0.0
  contact:
    name: 华电集团广东电厂监控平台研发组
servers:
  - url: http://localhost:8080
    description: 本地开发环境（Gateway 入口）
  - url: https://api.gd-energy.example.com
    description: 生产环境
security:
  - BearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [认证授权]
      summary: 用户登录
      description: 用户名密码登录，返回 JWT Token
      security: []
      parameters:
        - name: username
          in: query
          required: true
          description: 用户名
          schema:
            type: string
        - name: password
          in: query
          required: true
          description: 密码
          schema:
            type: string
            format: password
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginVO'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/R'
...
  /api/auth/refreshToken:
    post:
      tags: [认证授权]
      summary: 刷新Token
      description: 使用当前 Token 获取新的 JWT Token（Authorization 可选）
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginVO'
        '402':
          description: Token 已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/R'

  # ============================================================
  # 用户管理 (UserController) - 5 endpoints
  # ============================================================
...
  /api/plants:
    get:
      tags: [电厂管理]
      summary: 电厂列表（分页查询）
      description: 支持按类型、状态、关键字分页查询电厂列表
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: type
          in: query
          description: 电厂类型
          schema:
            $ref: '#/components/schemas/PlantType'
        - name: status
          in: query
          description: 运行状态
          schema:
            $ref: '#/components/schemas/PlantStatus'
        - name: keyword
          in: query
          description: 名称关键字（模糊查询）
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResultPlant'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [电厂管理]
      summary: 新建电厂
      description: 创建一个新电厂
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Plant'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Plant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/plants/nearby:
    get:
      tags: [电厂管理]
      summary: 附近电厂
      description: 根据经纬度和半径查询附近电厂
      parameters:
        - name: lng
          in: query
          required: true
          description: 经度
          schema:
            type: number
            format: double
        - name: lat
          in: query
          required: true
          description: 纬度
          schema:
            type: number
            format: double
        - name: radius
          in: query
          required: true
          description: 搜索半径（公里）
          schema:
            type: number
            format: double
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Plant'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/plants/{id}:
    get:
      tags: [电厂管理]
      summary: 电厂详情
      description: 根据电厂ID查询详情
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Plant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [电厂管理]
      summary: 更新电厂
      description: 更新电厂信息
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Plant'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Plant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [电厂管理]
      summary: 删除电厂
      description: 根据电厂ID删除电厂
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/R'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # 水文站管理 (HydroStationController) - 5 endpoints
  # ============================================================
  /api/hydro/stations:
    get:
      tags: [水文站管理]
      summary: 水文站列表（分页）
      description: 查询水文站列表，支持按城市、河流、关键字分页筛选
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: city
          in: query
          description: 所属城市
          schema:
            type: string
        - name: river
          in: query
          description: 所属河流
          schema:
            type: string
        - name: keyword
          in: query
          description: 名称关键字
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResultHydroStation'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [水文站管理]
      summary: 新建水文站
      description: 创建新水文站
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HydroStation'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HydroStation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/hydro/stations/{id}:
    get:
      tags: [水文站管理]
      summary: 水文站详情
      description: 根据水文站ID查询详情
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HydroStation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/hydro/stations/by-city/{city}:
    get:
      tags: [水文站管理]
      summary: 按城市查询水文站
      description: 根据城市名称查询该城市下所有水文站
      parameters:
        - name: city
          in: path
          required: true
          description: 城市名称
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HydroStation'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/hydro/stations/by-river/{river}:
    get:
      tags: [水文站管理]
      summary: 按河流查询水文站
      description: 根据河流名称查询该河流下所有水文站
      parameters:
        - name: river
          in: path
          required: true
          description: 河流名称
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/R'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HydroStation'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================
  # GIS 空间服务 (GisController) - 3 endpoints
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer Token 认证，登录后获取 token，请求头携带 Authorization: Bearer <token>"
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码（从1开始）
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 每页条数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdParam:
      name: id
      in: path
      required: true
      description: 主键ID
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: 请求参数错误（400 PARAM_ERROR）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
    Unauthorized:
      description: 未认证或Token无效（401 UNAUTHORIZED）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
...
    Forbidden:
      description: 无权限访问（403 FORBIDDEN）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
    NotFound:
      description: 资源不存在（404 NOT_FOUND）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/R'
...
  schemas:
...
    R:
      type: object
      description: |
        统一响应包装结构 R<T>。
        错误码：200 SUCCESS, 400 PARAM_ERROR, 401 UNAUTHORIZED,
        402 TOKEN_EXPIRED, 403 FORBIDDEN, 404 NOT_FOUND,
        500 BUSINESS_ERROR, 5000 SYSTEM_ERROR
      properties:
        code:
          type: integer
          description: 业务状态码
          example: 200
        msg:
          type: string
          description: 提示信息
          example: 操作成功
        data:
          description: 业务数据（具体结构见各接口）
          nullable: true

    # --------------------------------------------------------
    # 分页结果 PageResult<T>
    # --------------------------------------------------------
    PageResult:
      type: object
      description: 分页结果通用结构 PageResult<T>
      properties:
        list:
          type: array
          description: 数据列表
          items: {}
        total:
          type: integer
          format: int64
          description: 总记录数
        page:
          type: integer
          description: 当前页码
        size:
          type: integer
          description: 每页条数
...
    PageResultPlant:
      allOf:
        - $ref: '#/components/schemas/PageResult'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/Plant'
    PageResultHydroStation:
      allOf:
        - $ref: '#/components/schemas/PageResult'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/HydroStation'
...
    LoginVO:
      type: object
      description: 登录返回视图对象
      properties:
        token:
          type: string
          description: JWT Token
        userId:
          type: integer
          format: int64
          description: 用户ID
        username:
          type: string
          description: 用户名
        realName:
          type: string
          description: 真实姓名
        roles:
          type: array
          description: 角色列表
          items:
            $ref: '#/components/schemas/UserRole'
        permissions:
          type: array
          description: 权限标识列表
          items:
            type: string
...
    Plant:
      type: object
      description: 电厂信息
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        name:
          type: string
          description: 电厂名称
          example: 华电广州燃煤电厂
        type:
          $ref: '#/components/schemas/PlantType'
        lng:
          type: number
          format: double
          description: 经度
          example: 113.264385
        lat:
          type: number
          format: double
          description: 纬度
          example: 23.129112
        capacity:
          type: number
          format: double
          description: 装机容量（MW）
          example: 600.00
        status:
          $ref: '#/components/schemas/PlantStatus'
        address:
          type: string
          description: 详细地址
        createTime:
          type: string
          format: date-time
          description: 创建时间
        updateTime:
          type: string
          format: date-time
          description: 更新时间

    # --------------------------------------------------------
    # 水文站
    # --------------------------------------------------------
    HydroStation:
      type: object
      description: 水文站信息
      properties:
        id:
          type: integer
          format: int64
          description: 主键
        name:
          type: string
          description: 水文站名称
          example: 北江石角水文站
        riverBasin:
          type: string
          description: 所属流域/河流
          example: 北江
        lng:
          type: number
          format: double
          description: 经度
        lat:
          type: number
          format: double
          description: 纬度
        warningLevel:
          type: number
          format: double
          description: 警戒水位（m）
        guaranteeLevel:
          type: number
          format: double
          description: 保证水位（m）
        historicalMax:
          type: number
          format: double
          description: 历史最高水位（m）
        flowCapacity:
          type: number
          format: double
          description: 流量容量（m³/s）
        city:
          type: string
          description: 所属城市
          example: 清远
        status:
          type: integer
          description: 状态
        createTime:
          type: string
          format: date-time
          description: 创建时间
        updateTime:
          type: string
          format: date-time
          description: 更新时间

    # --------------------------------------------------------
    # 报警
    # --------------------------------------------------------
...
    UserRole:
      type: string
      description: 用户角色
      enum:
        - ADMIN
        - PROD_SAFETY
        - PLANT_MANAGER
        - OPERATOR
    PlantType:
      type: string
      description: 电厂类型
      enum:
        - coal
        - gas
        - solar
        - wind
        - storage
    PlantStatus:
      type: string
      description: 电厂运行状态
      enum:
        - normal
        - warning
        - danger
        - offline
...